+++ NEWS.md
@@ -1,7 +1,8 @@
 
-# rquery 0.4.1 2018/03/11
+# rquery 0.4.1 2018/03/13
 
  * Pipe database connection.
+ * Better quantile (assume window functions).
 
 # rquery 0.4.0 2018/03/10